fix: clean up was not quite right

+++ src/actions/cms/articles/destroy.ts
@@ -6,14 +6,28 @@ import { destroyArticleSection } from '../articleSections/destroy'
 
 export async function destroyArticle(id: number) : Promise<ActionReturn<Article>> {
     try {
-        const article = await prisma.article.delete({
+        const article = await prisma.article.findUnique({
             where: { id },
             include: { articleSections: true }
         })
+        if (!article) return { success: false, error: [{message: 'Article not found'}] }
+
+
         // destroy all articlesections in article
         for (const articleSection of article.articleSections) {
             await destroyArticleSection(articleSection.name)
         }
+        
+        // destroy coverimage
+        if (article.coverImageId) {
+            await prisma.cmsImage.delete({
+                where: { id: article.coverImageId }
+            })
+        }
+
+        await prisma.article.delete({
+            where: { id }
+        })
 
         return { 
             success: true,
